fix(enricher): handle empty result without exceptions (#160)
Before #153 enricher strategies used to throw a
`NodeUnsupportedException` in case they were tasked to enrich an unknown
node. Now they can simply return an empty list for improved performance.

With the old behavior: `len(results) == 0` <=> `len(exceptions) > 0`   
But now it's possible to get no results without exceptions in the
enricher.

All exceptions from the strategies were grouped in an `ExceptionGroup`
but the constructor of `ExceptionGroup` does not allow an empty list of
exceptions.

To fix this: We now through a different exception to handle this
edge-case.