fix: tests on main (#171)
Followup for #162

We added integration tests in #162 that did not account for the new
qubit output introduced in #170.
The pipeline in #162 did not fail, because it completed earlier (without
the new tests).

Our branch protection now enforces all branches to be up-to-date before
merging to prevent this in the future.